requirements.txt
pycryptodome==3.20.0
Jinja2==3.1.3
